refactor(subtitle): extract shared style controls and drop duplicate form
Hoist the full subtitle style editing form into a reusable
SubtitleStyleControls composable consumed by both the player bottom
sheet and the standalone tester, replacing the tester's minimal
StyleControls. The form is presentation-agnostic via showOverrideToggle
and onReset parameters so each host keeps its own chrome.

Also force MATCH_PARENT on preview surfaces (libVLC measured to zero
height and dropped its subtitle surface), select default mpv side-loaded
tracks instead of relying on auto heuristics, and wire a SAF font picker
into the tester via FontProvider.
