+++ feature/subtitle-tester/src/main/java/com/raulshma/jellyplay/feature/subtitle/tester/preview/PreviewEngineHost.kt
@@ -23,7 +23,17 @@ class PreviewEngineHost(
     fun attach(engine: MediaEngine) {
         detach()
         val surface = engine.createSurfaceView(context)
-        container.addView(surface)
+        // Force MATCH_PARENT: engines whose surface view does not set its own
+        // layout params (e.g. libVLC's VLCVideoLayout) would otherwise measure
+        // to wrap/zero content and render smaller than the preview tile. ExoPlayer
+        // and mpv set MATCH_PARENT internally, so this is a no-op for them.
+        container.addView(
+            surface,
+            FrameLayout.LayoutParams(
+                FrameLayout.LayoutParams.MATCH_PARENT,
+                FrameLayout.LayoutParams.MATCH_PARENT,
+            ),
+        )
         onSurfaceReady(surface)
     }
